# Optimize summing of losses in run_step() & Fix typos